requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
aiogram==3.0.0
SQLAlchemy==2.0.0
asyncpg==0.27.0
pydantic>=2.1.1
APScheduler==3.9.1
python-multipart==0.0.5
python-dotenv
requests
httpx